Extract useTemplates to its own file
