+++ packages/edit-site/src/components/left-sidebar/navigation-panel/menus/templates.js
@@ -1,18 +1,16 @@
 /**
  * WordPress dependencies
  */
-import { useEffect, useState } from '@wordpress/element';
-import { useSelect } from '@wordpress/data';
 import {
 	__experimentalNavigationItem as NavigationItem,
 	__experimentalNavigationMenu as NavigationMenu,
 } from '@wordpress/components';
-import apiFetch from '@wordpress/api-fetch';
 
 /**
  * Internal dependencies
  */
 import TemplatesPagesMenu from './templates-pages';
+import useTemplates from '../use-templates';
 
 const GENERAL_TEMPLATE_SLUGS = [
 	'front-page',
@@ -23,38 +21,6 @@ const GENERAL_TEMPLATE_SLUGS = [
 	'search',
 	'404',
 ];
-
-function useTemplates() {
-	const [ templateFiles, setTemplateFiles ] = useState( [] );
-
-	useEffect( () => {
-		apiFetch( {
-			path: '/__experimental/edit-site/v1/page-templates',
-		} ).then( ( res ) => {
-			setTemplateFiles( res );
-		} );
-	}, [] );
-
-	const templateEntities = useSelect(
-		( select ) =>
-			select( 'core' ).getEntityRecords( 'postType', 'wp_template' ) ||
-			[],
-		[]
-	);
-
-	const templatesFilesWithoutEntities = templateFiles.filter(
-		( slug ) =>
-			! templateEntities?.find( ( entity ) => slug === entity.slug )
-	);
-
-	return [
-		...templatesFilesWithoutEntities.map( ( slug ) => ( {
-			id: slug,
-			slug,
-		} ) ),
-		...templateEntities,
-	];
-}
 
 export default function TemplatesMenu( { onActiveIdChange } ) {
 	const templates = useTemplates();
